Move file and fix imports
